S09P22A103-310 fix: db 자동생성 수정 db 자동생성 수정

+++ backend/daengdaeng/src/main/java/com/daengdaeng/domain/category/domain/Category.java
@@ -20,7 +20,7 @@ import lombok.NoArgsConstructor;
 public class Category {
 
 	@Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int categoryId;
+	private byte categoryId;
 
 	@Column(length = 20, unique = true, nullable = false)
 	private String category;
